Rename file and class
Implements FileMaster class which is supposed to handle loading of scenarios and saves, as well as saving and loading of game state to disk.
Loading/saving functionality might be split into their own classes in the future to separate functionality, but the FileMaster should still be the interface to accomplish these objectives.