Rework hexmap for easier use in renderer
Moved method to translate from cubic to pixel space into the Axial class.
Changed the name of the Map class to TileMap.
Removed all references to the size of the tiles or grid out of the module. This has to be handled in renderer modules.
Reworked TileMap to allow for other objects than Hexes. Hexes will be removed when all functionality is moved to the Entity.tile.Tile class.
Reformatted TileMap so relevant methods are grouped together.